// frontend/src/api/models/handlerInitiateUploadRequest.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */
import type { HandlerInitiateUploadRequestType } from './handlerInitiateUploadRequestType'

/**
 * Request body for initiating a file upload
 */
export interface HandlerInitiateUploadRequest {
  content_type: string
  /**
   * @minLength 1
   * @maxLength 255
   */
  file_name: string
  /**
   * max 100MB
   * @maximum 104857600
   */
  file_size: number
  product_id: string
  type: HandlerInitiateUploadRequestType
}