// frontend/src/api/models/handlerInitiateUploadRequestType.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type HandlerInitiateUploadRequestType =
  (typeof HandlerInitiateUploadRequestType)[keyof typeof HandlerInitiateUploadRequestType]

export const HandlerInitiateUploadRequestType = {
  main_image: 'main_image',
  gallery_image: 'gallery_image',
  document: 'document',
  other: 'other',
} as const